chess_game/requirement.txt
pygame>=2.5.2 
python-chess==1.999